requirements.txt
aiogram
groq
python-dotenv
google-generativeai
aiohttp
requests
asyncpg